// src/app/models/survival-analysis/granularity-type.ts
/**
 * Copyright 2020 CHUV
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */


export enum Granularity {
  day = 'day',
  week = 'week',
  month = 'month',
  year = 'year',
}